+++ tests/conftest.py
@@ -8,6 +8,11 @@ def one_product():
 
 
 @pytest.fixture
+def second_product():
+    return Product("Samsung", "256GB, White", 40000.0, 2)
+
+
+@pytest.fixture
 def one_category(one_product):
     return Category("Смартфоны",
                     "Смартфоны, как средство не только коммуникации, "
