+++ tests/test_models.py
@@ -1,0 +1,40 @@
+from src.models import Product
+
+
+def test_product(one_product):
+    assert one_product.name == "Iphone 15"
+    assert one_product.description == "512GB, Gray space"
+    assert one_product.price == 210000.0
+    assert one_product.quantity == 8
+
+
+def test_category(one_category):
+    assert one_category.name == "Смартфоны"
+    assert one_category.description == ("Смартфоны, как средство не только коммуникации, "
+                                        "но и получения дополнительных функций для удобства жизни")
+    assert one_category.category_count == 1
+    assert one_category.product_count == 1
+
+
+def test_views_product(one_category):
+    assert one_category.views_product == "Iphone 15, 210000.0. Остаток: 8\n"
+
+
+def test_views_product_setter(one_category, second_product):
+    assert one_category.product_count == 3
+    one_category.views_product = second_product
+    assert one_category.product_count == 4
+
+
+def test_new_product():
+    product = Product("Nokia", "120GB, Gray space", 15000.0, 4)
+    assert product.name == "Nokia"
+
+
+def test_price_getter(one_product):
+    assert one_product.price == 210000
+
+
+def test_price_setter(capsys, one_product):
+    one_product.price = 300000
+    assert one_product.price == 300000
